backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow>=2.20.0
pillow>=11.0.0
numpy>=1.24.3
python-multipart==0.0.6
python-dotenv==1.0.0
openai
